requirements.txt
numpy==1.24.2
opencv_python==4.7.0.68
Pillow==9.5.0
PyYAML==6.0
shapely==2.0.1
